fix: move lab capability planning into core runner
