+++ src/commands/utils/resource_policy.rs
@@ -1,4 +1,3 @@
-use std::path::Path;
 use std::sync::RwLock;
 
 use serde::{Deserialize, Serialize};
@@ -6,7 +5,6 @@ use serde::{Deserialize, Serialize};
 use crate::cli_surface::Commands;
 
 use crate::commands::doctor::resources::{DoctorOutput, ResourceRecommendation};
-use crate::commands::runner::doctor::RunnerDoctorOutput;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct HotCommand {
@@ -38,82 +36,6 @@ pub struct ResourcePolicyWarning {
     pub command: &'static str,
     pub recommendation: ResourceRecommendation,
     pub message: String,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum LabRunnerGateMode {
-    Automatic,
-    Explicit,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum LabRunnerTool {
-    Git,
-    Node,
-    Npm,
-    Pnpm,
-    Php,
-    Composer,
-    Docker,
-    Playwright,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct LabRunnerCapabilityPlan {
-    pub command: &'static str,
-    pub required_tools: Vec<LabRunnerTool>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum LabRunnerGateDecision {
-    Eligible,
-    Missing {
-        runner_id: String,
-        command: &'static str,
-        missing_tools: Vec<LabRunnerTool>,
-        reason: String,
-        remediation: Vec<String>,
-    },
-}
-
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct LabRunnerCapabilities {
-    pub git: bool,
-    pub node: bool,
-    pub npm: bool,
-    pub pnpm: bool,
-    pub php: bool,
-    pub composer: bool,
-    pub docker: bool,
-    pub playwright: bool,
-}
-
-impl LabRunnerCapabilities {
-    pub fn from_doctor(report: &RunnerDoctorOutput) -> Self {
-        Self {
-            git: report.capabilities.git,
-            node: report.capabilities.node,
-            npm: report.capabilities.npm,
-            pnpm: report.capabilities.pnpm,
-            php: report.capabilities.php,
-            composer: report.capabilities.composer,
-            docker: report.capabilities.docker,
-            playwright: report.capabilities.playwright && report.capabilities.browser_ready,
-        }
-    }
-
-    fn has(&self, tool: LabRunnerTool) -> bool {
-        match tool {
-            LabRunnerTool::Git => self.git,
-            LabRunnerTool::Node => self.node,
-            LabRunnerTool::Npm => self.npm,
-            LabRunnerTool::Pnpm => self.pnpm,
-            LabRunnerTool::Php => self.php,
-            LabRunnerTool::Composer => self.composer,
-            LabRunnerTool::Docker => self.docker,
-            LabRunnerTool::Playwright => self.playwright,
-        }
-    }
 }
 
 /// Persisted, host-pressure context captured at preflight time for any
@@ -292,148 +214,6 @@ pub fn hot_command(command: &Commands) -> Option<HotCommand> {
     }
 }
 
-pub fn lab_runner_capability_plan(
-    command: &Commands,
-    source_path: &Path,
-) -> Option<LabRunnerCapabilityPlan> {
-    let hot_command = hot_command(command)?;
-    let mut required_tools = vec![LabRunnerTool::Git];
-
-    if source_path.join("package.json").is_file() {
-        push_unique(&mut required_tools, LabRunnerTool::Node);
-        push_unique(&mut required_tools, LabRunnerTool::Npm);
-    }
-
-    if source_path.join("pnpm-lock.yaml").is_file() {
-        push_unique(&mut required_tools, LabRunnerTool::Node);
-        push_unique(&mut required_tools, LabRunnerTool::Pnpm);
-    }
-
-    if source_path.join("composer.json").is_file() {
-        push_unique(&mut required_tools, LabRunnerTool::Php);
-        push_unique(&mut required_tools, LabRunnerTool::Composer);
-    }
-
-    if has_docker_signal(source_path) {
-        push_unique(&mut required_tools, LabRunnerTool::Docker);
-    }
-
-    if matches!(command, Commands::Trace(_)) {
-        push_unique(&mut required_tools, LabRunnerTool::Playwright);
-    }
-
-    Some(LabRunnerCapabilityPlan {
-        command: hot_command.label,
-        required_tools,
-    })
-}
-
-pub fn evaluate_lab_runner_capabilities(
-    runner_id: &str,
-    plan: &LabRunnerCapabilityPlan,
-    capabilities: &LabRunnerCapabilities,
-    mode: LabRunnerGateMode,
-) -> LabRunnerGateDecision {
-    let missing_tools = plan
-        .required_tools
-        .iter()
-        .copied()
-        .filter(|tool| !capabilities.has(*tool))
-        .collect::<Vec<_>>();
-
-    if missing_tools.is_empty() {
-        return LabRunnerGateDecision::Eligible;
-    }
-
-    let missing = tool_list(&missing_tools);
-    let reason = match mode {
-        LabRunnerGateMode::Automatic => format!(
-            "Local fallback: runner '{runner_id}' is missing required tool(s) for `{}`: {missing}.",
-            plan.command
-        ),
-        LabRunnerGateMode::Explicit => format!(
-            "Lab offload runner '{runner_id}' is missing required tool(s) for `{}`: {missing}.",
-            plan.command
-        ),
-    };
-    let mut remediation = missing_tools
-        .iter()
-        .map(|tool| tool.remediation().to_string())
-        .collect::<Vec<_>>();
-    match mode {
-        LabRunnerGateMode::Automatic => remediation.push(
-            "Homeboy will run locally instead of auto-offloading to this runner.".to_string(),
-        ),
-        LabRunnerGateMode::Explicit => remediation.push(
-            "Install the missing tool(s) on the runner, choose a capable runner, or omit --runner to run locally.".to_string(),
-        ),
-    }
-
-    LabRunnerGateDecision::Missing {
-        runner_id: runner_id.to_string(),
-        command: plan.command,
-        missing_tools,
-        reason,
-        remediation,
-    }
-}
-
-fn push_unique<T: PartialEq>(items: &mut Vec<T>, item: T) {
-    if !items.contains(&item) {
-        items.push(item);
-    }
-}
-
-fn has_docker_signal(source_path: &Path) -> bool {
-    [
-        "Dockerfile",
-        "docker-compose.yml",
-        "docker-compose.yaml",
-        "compose.yml",
-        "compose.yaml",
-    ]
-    .iter()
-    .any(|name| source_path.join(name).is_file())
-}
-
-fn tool_list(tools: &[LabRunnerTool]) -> String {
-    tools
-        .iter()
-        .map(|tool| tool.id())
-        .collect::<Vec<_>>()
-        .join(", ")
-}
-
-impl LabRunnerTool {
-    pub fn id(self) -> &'static str {
-        match self {
-            LabRunnerTool::Git => "git",
-            LabRunnerTool::Node => "node",
-            LabRunnerTool::Npm => "npm",
-            LabRunnerTool::Pnpm => "pnpm",
-            LabRunnerTool::Php => "php",
-            LabRunnerTool::Composer => "composer",
-            LabRunnerTool::Docker => "docker",
-            LabRunnerTool::Playwright => "playwright+browsers",
-        }
-    }
-
-    fn remediation(self) -> &'static str {
-        match self {
-            LabRunnerTool::Git => "Install git and ensure it is on the runner PATH.",
-            LabRunnerTool::Node => "Install Node.js and ensure node is on the runner PATH.",
-            LabRunnerTool::Npm => "Install npm with Node.js and ensure npm is on the runner PATH.",
-            LabRunnerTool::Pnpm => "Install pnpm for repositories with pnpm-lock.yaml.",
-            LabRunnerTool::Php => "Install PHP for repositories with composer.json.",
-            LabRunnerTool::Composer => "Install Composer for repositories with composer.json.",
-            LabRunnerTool::Docker => "Install and start Docker for container-backed repositories.",
-            LabRunnerTool::Playwright => {
-                "Install Playwright CLI and browser binaries on the runner."
-            }
-        }
-    }
-}
-
 pub fn evaluate(command: HotCommand, resources: &DoctorOutput) -> Option<ResourcePolicyWarning> {
     match resources.recommendation {
         ResourceRecommendation::Ok => None,
@@ -519,36 +299,6 @@ mod tests {
     use crate::commands::doctor::resources::{
         LoadSummary, MemorySummary, ProcessSummary, RigLeaseSummary,
     };
-    use crate::commands::lint::LintArgs;
-    use crate::commands::utils::args::{
-        BaselineArgs, ExtensionOverrideArgs, PositionalComponentArgs, SettingArgs,
-    };
-
-    fn lint_command() -> Commands {
-        Commands::Lint(LintArgs {
-            comp: PositionalComponentArgs {
-                component: None,
-                path: None,
-            },
-            extension_override: ExtensionOverrideArgs::default(),
-            summary: false,
-            file: None,
-            glob: None,
-            changed_only: false,
-            changed_since: None,
-            ci_job: None,
-            errors_only: false,
-            sniffs: None,
-            exclude_sniffs: None,
-            category: None,
-            fix: false,
-            force: false,
-            setting_args: SettingArgs::default(),
-            baseline_args: BaselineArgs::default(),
-            json_summary: false,
-        })
-    }
-
     fn resources(recommendation: ResourceRecommendation) -> DoctorOutput {
         DoctorOutput {
             command: "doctor.resources",
@@ -738,121 +488,5 @@ mod tests {
         assert!(value["message"].is_string());
         assert_eq!(value["host"]["load_severity"], "hot");
         assert_eq!(value["host"]["cpu_count"], 4);
-    }
-
-    #[test]
-    fn lab_runner_capability_plan_detects_project_tool_needs() {
-        let dir = tempfile::tempdir().expect("temp project");
-        std::fs::write(dir.path().join("package.json"), "{}").expect("package.json");
-        std::fs::write(dir.path().join("pnpm-lock.yaml"), "lockfileVersion: '9'\n")
-            .expect("pnpm-lock");
-        std::fs::write(dir.path().join("composer.json"), "{}").expect("composer.json");
-        std::fs::write(dir.path().join("Dockerfile"), "FROM scratch\n").expect("Dockerfile");
-
-        let plan = lab_runner_capability_plan(&lint_command(), dir.path()).expect("plan");
-
-        assert_eq!(plan.command, "lint");
-        assert_eq!(
-            plan.required_tools,
-            vec![
-                LabRunnerTool::Git,
-                LabRunnerTool::Node,
-                LabRunnerTool::Npm,
-                LabRunnerTool::Pnpm,
-                LabRunnerTool::Php,
-                LabRunnerTool::Composer,
-                LabRunnerTool::Docker,
-            ]
-        );
-    }
-
-    #[test]
-    fn lab_runner_gate_allows_capable_runner() {
-        let plan = LabRunnerCapabilityPlan {
-            command: "lint",
-            required_tools: vec![LabRunnerTool::Git, LabRunnerTool::Node, LabRunnerTool::Pnpm],
-        };
-        let capabilities = LabRunnerCapabilities {
-            git: true,
-            node: true,
-            pnpm: true,
-            ..LabRunnerCapabilities::default()
-        };
-
-        assert_eq!(
-            evaluate_lab_runner_capabilities(
-                "lab",
-                &plan,
-                &capabilities,
-                LabRunnerGateMode::Explicit,
-            ),
-            LabRunnerGateDecision::Eligible
-        );
-    }
-
-    #[test]
-    fn lab_runner_gate_reports_missing_tool_for_explicit_runner() {
-        let plan = LabRunnerCapabilityPlan {
-            command: "test",
-            required_tools: vec![LabRunnerTool::Git, LabRunnerTool::Pnpm],
-        };
-        let capabilities = LabRunnerCapabilities {
-            git: true,
-            ..LabRunnerCapabilities::default()
-        };
-
-        let decision = evaluate_lab_runner_capabilities(
-            "lab",
-            &plan,
-            &capabilities,
-            LabRunnerGateMode::Explicit,
-        );
-
-        let LabRunnerGateDecision::Missing {
-            missing_tools,
-            reason,
-            remediation,
-            ..
-        } = decision
-        else {
-            panic!("expected missing tool decision");
-        };
-        assert_eq!(missing_tools, vec![LabRunnerTool::Pnpm]);
-        assert!(reason.contains("Lab offload runner 'lab'"));
-        assert!(reason.contains("pnpm"));
-        assert!(remediation
-            .iter()
-            .any(|item| item.contains("omit --runner")));
-    }
-
-    #[test]
-    fn lab_runner_gate_reports_local_fallback_for_auto_runner() {
-        let plan = LabRunnerCapabilityPlan {
-            command: "trace",
-            required_tools: vec![LabRunnerTool::Git, LabRunnerTool::Playwright],
-        };
-        let capabilities = LabRunnerCapabilities {
-            git: true,
-            ..LabRunnerCapabilities::default()
-        };
-
-        let decision = evaluate_lab_runner_capabilities(
-            "lab",
-            &plan,
-            &capabilities,
-            LabRunnerGateMode::Automatic,
-        );
-
-        let LabRunnerGateDecision::Missing {
-            reason,
-            remediation,
-            ..
-        } = decision
-        else {
-            panic!("expected local fallback decision");
-        };
-        assert!(reason.contains("Local fallback"));
-        assert!(reason.contains("playwright+browsers"));
-        assert!(remediation.iter().any(|item| item.contains("run locally")));
     }
 }
